+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - ICollisionable interface
 - BoxInteractor component
+- SphereInteractor component
 
 ## [1.0.0] - 2025-10-30
 ### Added